linkify comment text
