+++ modules/periodo-ui/src/index.js
@@ -43,4 +43,5 @@ addUIModules([
   require('./Typography'),
   require('./WorldMap'),
   require('./diffable/Value'),
+  require('./util'),
 ])
